minh

+++ src/main/java/nhom8/controller/BillController.java
@@ -58,6 +58,7 @@ public class BillController implements ManagerController {
         try {
             Bill bill = new Bill();
             bill.setStatus(false);
+            bill.setUserId(view.getUser().getId());
             Map<String, Object> result = billDAO.create(bill);
             if ((boolean) result.get("status")) {
                 this.bill = billDAO.getNewBill();
@@ -133,19 +134,23 @@ public class BillController implements ManagerController {
     @Override
     public void actionDelete() {
         jdDelete.getBtnDelete().addActionListener(e -> {
-            try {
-                Map<String, Object> result = billDAO.delete(bill.getId());
-                if ((Boolean) result.get("status")) {
-                    JOptionPane.showMessageDialog(jdDelete, result.get("message"));
-                    System.out.println(result.get("message"));
-                    updateData();
-                    jdDelete.dispose();
-                } else {
-                    JOptionPane.showMessageDialog(jdDelete, result.get("message"));
+            if (bill.getUserId().equals(view.getUser().getId()) || view.getUser().getRole() == 1) {
+                try {
+                    Map<String, Object> result = billDAO.delete(bill.getId());
+                    if ((Boolean) result.get("status")) {
+                        JOptionPane.showMessageDialog(jdDelete, result.get("message"));
+                        System.out.println(result.get("message"));
+                        updateData();
+                    } else {
+                        JOptionPane.showMessageDialog(jdDelete, result.get("message"));
+                    }
+                } catch (SQLException ex) {
+                    throw new RuntimeException(ex);
                 }
-            } catch (SQLException ex) {
-                throw new RuntimeException(ex);
+            } else {
+                JOptionPane.showMessageDialog(jdDelete, "Ban chi duoc phep xoa don ban tao");
             }
+            jdDelete.dispose();
         });
     }
 
@@ -208,11 +213,11 @@ public class BillController implements ManagerController {
             bills = billDAO.getAll();
             panel.getTblBill().removeAll();
 
-            String[] cols = {"Id", "Trang thai", "Tong tien", "Ngay tao", "Ngay cap nhat"};
+            String[] cols = {"Id", "Nguoi tao", "Trang thai", "Tong tien", "Ngay tao", "Ngay cap nhat"};
             DefaultTableModel dtm = new DefaultTableModel(cols, 0);
             if (!Common.isNullOrEmpty(bills)) {
                 bills.forEach(obj -> {
-                    dtm.addRow(new Object[]{obj.getId(), obj.getStatus() ? "Da thanh toan" : "Chua thanh toan", Common.isNullOrEmpty(obj.getTotal()) ? 0 : obj.getTotal(), obj.getCreatedAt(),
+                    dtm.addRow(new Object[]{obj.getId(), obj.getUserName(), obj.getStatus() ? "Da thanh toan" : "Chua thanh toan", Common.isNullOrEmpty(obj.getTotal()) ? 0 : obj.getTotal(), obj.getCreatedAt(),
                             obj.getUpdatedAt()});
                 });
